// src/foxapi/types/evalcontext.ts
export type EvalContext =
  | "ADD_COMMAND"
  | "ADD_GIVEAWAY"
  | "ADD_POLL"
  | "ADD_PERMISSION"
  | "ADD_TICKET"
  | "DELETE_COMMAND"
  | "DELETE_GIVEAWAY"
  | "DELETE_PERMISSION"
  | "DELETE_POLL"
  | "GET_PACKAGES"
  | "GET_POLLS"
  | "GET_SETTINGS"
  | "GET_PERMISSIONS"
  | "GET_GIVEAWAYS"
  | "GET_LEVELING"
  | "GET_TICKETS"
  | "GET_CHANNELS"
  | "GET_ROLES"
  | "GET_EMOJIS"
  | "GET_BANKING"
  | "GET_COMMANDS"
  | "PATCH_COMMAND"
  | "PATCH_GIVEAWAY"
  | "PATCH_POLL"
  | "PATCH_LEVELING"
  | "PATCH_BANKING"
  | "TOGGLE_SETTING"
  | "TOGGLE_SETTING_ARRAY"
  | "TOGGLE_PACKAGE";